% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dri_ic.R
\name{dri_plot}
\alias{dri_plot}
\title{Plot DRI IC}
\usage{
dri_plot(data, x, y, title = NA, suffix = NA, DRI = NA)
}
\arguments{
\item{data}{IC}

\item{x}{name of x variable}

\item{y}{name of y variable}

\item{title}{title of the plot (default = NA)}

\item{suffix}{string pre, post (default = NA)}

\item{DRI}{numeric value (default = NA)}
}
\value{
plot
}
\description{
Plot DRI IC
}
\examples{
dri_plot(data, x, y)
}
