% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dri_survey.R
\name{format_dri_survey}
\alias{format_dri_survey}
\title{Formats surveys for further processing.}
\usage{
format_dri_survey(
  survey_info = list(type = NA_character_, order = NA_integer_, statement =
    NA_character_, name = NA_character_, scale_max = NA_integer_, q_method = NA, )
)
}
\arguments{
\item{survey_info}{survey information needed to format dri survey.}
}
\value{
A list of survey info.
}
\description{
Formats surveys for further processing.
}
\examples{
\dontrun{
dri_survey <- format_dri_survey(deliberr::surveys \%>\% filter(name == "acp"))
}
}
