% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri.R
\name{get_dri}
\alias{get_dri}
\title{Get DRI (group)}
\usage{
get_dri(ic, adjusted = TRUE)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}

\item{adjusted}{boolean flag indicates whether to use updated or original DRI
formula (default = TRUE)}
}
\value{
the group-level DRI value
}
\description{
Get DRI (group)
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

ic <- get_dri_ic(data)
get_dri(ic)

}
