% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_alpha.R
\name{get_dri_alpha}
\alias{get_dri_alpha}
\title{Get DRI Cronbach's Alpha}
\usage{
get_dri_alpha(data)
}
\arguments{
\item{data}{the raw survey data}
}
\value{
a dataframe with alpha_c, alpha_p, and alpha_all with values of
Cronbach's alpha for considerations, policy preferences, and both, respectively.
}
\description{
Get DRI Cronbach's Alpha
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

get_dri_alpha(data)

}
