% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_case.R
\name{get_dri_case}
\alias{get_dri_case}
\title{Get DRI Case}
\usage{
get_dri_case(case, adjusted = TRUE, method = "wilcox", alternative = "greater")
}
\arguments{
\item{case}{a character string specifying the name of the case in human_data}

\item{adjusted}{a logical indicating whether you want the original or adjusted
DRI formula}

\item{method}{a character string specifying the method for statistical testing,
must be one of "wilcox" (default) or "t.test"}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "greater" (default), "two.sided" or "less". You can specify just the initial letter.}
}
\value{
a tibble with with the following components: \code{case}, \code{pre},
\code{post}, \code{delta}, \code{p_value}, and \code{significance}.
}
\description{
Get the DRI of the specified case in human data
}
\examples{

case <- "Citizen Parliamentarian"
get_dri_case("Activate")

}
