% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_ic.R
\name{get_dri_ic}
\alias{get_dri_ic}
\title{get DRI IC}
\usage{
get_dri_ic(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
get DRI IC
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)
get_dri_ic(data)
}
