% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_ind.R
\name{get_dri_ind}
\alias{get_dri_ind}
\title{Get DRI (individual)}
\usage{
get_dri_ind(ic)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}
}
\value{
dataframe with individual DRI
}
\description{
Get DRI (individual)
}
\examples{
get_dri_ind(ic)
}
