% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_llm_response.R
\name{get_dri_llm_response}
\alias{get_dri_llm_response}
\title{Get DRI LLM response}
\usage{
get_dri_llm_response(model_id, survey_name, role_uid = NA_character_)
}
\arguments{
\item{model_id}{The model_id from openrounter.ai}

\item{survey_name}{The name of the survey}

\item{role_uid}{The unique identifier of a role (optional)}
}
\value{
A data frame with one row of data.
}
\description{
Uses openrouter.ai to generate LLM responses to DRI survey questions.
}
\examples{

model_id <- "google/gemini-2.5-flash-lite"
survey_name <- "ccps"
get_dri_llm_response(model_id, survey_name, "eco")
}
