% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_survey_info.R
\name{get_dri_survey_info}
\alias{get_dri_survey_info}
\title{Get survey info}
\usage{
get_dri_survey_info(survey_name = NA_character_)
}
\arguments{
\item{survey_name}{the name of the survey (optional).}
}
\value{
A list of survey info.
}
\description{
Get survey info
}
\examples{
survey_info <- get_dri_survey_info("acp")
survey_info[[1]]$name
survey_info[[1]]$considerations
length(survey_info[[1]]$considerations$statement)
survey_info[[1]]$policies
length(survey_info[[1]]$policies$statement)
survey_info[[1]]$scale_max
survey_info[[1]]$q_method
}
