% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_ids.R
\name{get_model_ids}
\alias{get_model_ids}
\title{Get model ids}
\usage{
get_model_ids()
}
\value{
A dataframe with columns \code{provider} and \code{model}
}
\description{
Gets the provider and model names from openrouter. The \code{model_id} can
be recreated as \code{provider}/\code{model}
}
\examples{

get_model_ids()

}
