% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_data}
\alias{human_data}
\title{Human data}
\format{
A data frame with 67 variables, including
\code{survey}, \code{case}, \code{stage_id}, \code{C1...C50} and \code{P1...P10}.
}
\usage{
human_data
}
\description{
Pre- and post-deliberation DRI survey data from 24 deliberation cases around
the world. Some cases used the same survey.
}
\keyword{datasets}
