% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dri_llm_prompts.R
\name{make_dri_llm_prompts}
\alias{make_dri_llm_prompts}
\title{Make DRI LLM prompts}
\usage{
make_dri_llm_prompts(
  dri_survey,
  role_info = list(uid = NA_character_, role = NA_character_, description =
    NA_character_)
)
}
\arguments{
\item{dri_survey}{A list generated using format_dri_survey(...).}

\item{role_info}{Information about a specific role.}
}
\value{
A list of lists with four variables: \code{system},
\code{considerations}, \code{policies}, and \code{reason}.
}
\description{
Creates the prompts for generating LLM DRI data.
}
\examples{
dri_survey <- format_dri_survey(deliberr::surveys[surveys$name == "ccps",])
role_info <- list(
  uid = "sur",
  role = "surfer",
  description = "likes the ocean"
)
make_dri_llm_prompts(dri_survey, role_info)
}
