% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_dri.R
\name{permute_dri}
\alias{permute_dri}
\title{Tests whether the links between considerations and policy preferences are
consistent or due to chance.}
\usage{
permute_dri(data, iterations = 10000, verbose = FALSE, summary = TRUE)
}
\arguments{
\item{data}{raw dri survey dataframe}

\item{iterations}{number permutations to generate. Default = 10000}

\item{verbose}{boolean flag to print time of permutation. Default = FALSE}

\item{summary}{whether to return the raw data or summary of test results}
}
\value{
dataframe with permutation test results, raw or summarized
}
\description{
Tests whether the links between considerations and policy preferences are
consistent or due to chance.
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

permute_dri(data, iterations = 100)

}
