% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prompts}
\alias{prompts}
\title{Prompts}
\format{
A data frame with two variables:
\code{type} and \code{prompt}
}
\usage{
prompts
}
\description{
Prompts used in DRI surveys for humans and LLMs.
}
\keyword{datasets}
