% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roles}
\alias{roles}
\title{Roles}
\format{
A data frame with five variables:
\code{uid}, \code{type}, \code{article}, \code{role} and \code{description}.
}
\usage{
roles
}
\description{
Roles used to generate LLM role-playing data.
}
\keyword{datasets}
