% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_perm_dri.R
\name{summarize_perm_dri}
\alias{summarize_perm_dri}
\title{Summarize the results of a permutation test}
\usage{
summarize_perm_dri(perms, type = "common")
}
\arguments{
\item{perms}{results of the permutation test}

\item{type}{which type of statistics to summarize (e.g., common, robust, mean)}
}
\value{
summary of permutation test
}
\description{
Summarize the results of a permutation test
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

perms <- permute_dri(data, iterations = 100, summary = FALSE)
summarize_perm_dri(perms)

}
\seealso{
\code{\link[rstatix:get_summary_stats]{rstatix::get_summary_stats()}} for values of "type".
}
