% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Surveys}
\format{
A data frame with six variables:
\code{type}, \code{order}, \code{statement}, \code{name}, \code{scale_max} and \code{q_method}.
}
\usage{
surveys
}
\description{
All survey data used in deliberations.
}
\keyword{datasets}
