% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliberr-package.R
\docType{package}
\name{deliberr-package}
\alias{deliberr}
\alias{deliberr-package}
\title{deliberr: Deliberation in R}
\description{
This package includes functions for analyzing deliberation data, including Deliberative Reasoning Index (DRI).
}
\author{
\strong{Maintainer}: Gustavo Kreia Umbelino \email{gumbelin@gmail.com}

}
\keyword{internal}
