% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri.R
\name{get_dri}
\alias{get_dri}
\title{Get DRI (group)}
\usage{
get_dri(ic, adjusted = TRUE)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}

\item{adjusted}{boolean flag indicates whether to use updated or original DRI
formula (default = TRUE)}
}
\value{
the group-level DRI value
}
\description{
Get DRI (group)
}
\examples{
get_dri(ic)
}
