% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_alpha.R
\name{get_dri_alpha}
\alias{get_dri_alpha}
\title{Get DRI Cronbach's Alpha}
\usage{
get_dri_alpha(data)
}
\arguments{
\item{data}{the raw survey data}
}
\value{
a dataframe with alpha_c, alpha_p, and alpha_all with values of
Cronbach's alpha for considerations, policy preferences, and both, respectively.
}
\description{
Get DRI Cronbach's Alpha
}
\examples{
get_dri_alpha(data)
}
