% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dri_ind.R
\name{get_dri_ind}
\alias{get_dri_ind}
\title{Get DRI (individual)}
\usage{
get_dri_ind(ic, adjusted = TRUE)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}

\item{adjusted}{a logical indicating whether you want the original or adjusted
DRI formula}
}
\value{
dataframe with individual DRI
}
\description{
Get DRI (individual)
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

ic <- get_dri_ic(data)
get_dri_ind(ic)
}
