% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dri_llm_prompts.R
\name{make_dri_llm_prompts}
\alias{make_dri_llm_prompts}
\title{Make DRI LLM prompts}
\usage{
make_dri_llm_prompts(survey_info, role_uid = NA_character_)
}
\arguments{
\item{survey_info}{A list generated using get_dri_survey_info().}

\item{role_uid}{The unique identifier of a role (optional).}
}
\value{
A list of lists with four variables: \code{system},
\code{considerations}, \code{policies}, and \code{reason}.
}
\description{
Creates the prompts for generating LLM DRI data.
}
\examples{
survey_info <- get_dri_survey_info("ccps")
make_dri_llm_prompts(survey_info[[1]], "csk")
}
