% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dri_ic.R
\name{plot_dri_ic}
\alias{plot_dri_ic}
\title{Plot DRI IC}
\usage{
plot_dri_ic(ic, title = NA, suffix = NA, DRI = NA)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}

\item{title}{title of the plot (default = NA)}

\item{suffix}{string pre, post (default = NA)}

\item{DRI}{numeric value (default = NA)}
}
\value{
plot
}
\description{
Plot DRI IC
}
\examples{
plot_dri_ic(ic)
}
