% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dri_ic.R
\name{plot_dri_ic}
\alias{plot_dri_ic}
\title{Plot DRI IC}
\usage{
plot_dri_ic(ic, title = NA, suffix = NA, dri = NA)
}
\arguments{
\item{ic}{dataframe generated by get_dri_ic(data)}

\item{title}{title of the plot (default = NA)}

\item{suffix}{string pre, post (default = NA)}

\item{dri}{numeric value (default = NA)}
}
\value{
plot
}
\description{
Plot DRI IC
}
\examples{

library(tibble)
data <- tibble(
  pnum = c(1,2,3,4),
  C1 = c(1,2,3,4),
  C2 = c(2,3,4,3),
  C3 = c(3,2,4,2),
  C4 = c(4,3,3,1),
  P1 = c(1,2,3,3),
  P2 = c(2,3,1,2),
  P3 = c(3,1,2,1),
)

ic <- get_dri_ic(data)
plot_dri_ic(ic)
}
